+++ dsmr_mqtt/tests/services/test_broker.py
@@ -169,7 +169,7 @@ class TestBroker(TestCase):
 
         # We assert that the LAST X messages were sent, rest is deleted.
         for x in range(MAX + 1, MAX * 2 + 1):
-            publish_mock.assert_any_call(topic='z', payload=str(x), qos=0, retain=True)
+            publish_mock.assert_any_call(topic='z', payload=str(x), qos=2, retain=True)
 
         self.assertFalse(queue.Message.objects.exists())